pyproject.toml
[project]
name = "digital-empathy-platform"
version = "0.1.0"
description = "FastAPI backend and realtime services for the Digital Empathy Platform"
authors = [{name = "Digital Empathy Team", email = "team@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy[asyncio]>=2.0.29",
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tenacity>=8.2.3",
    "pyyaml>=6.0.1",
    "prometheus-client>=0.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0"
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
